+++ test/test_semantic_chunker.py
@@ -13,7 +13,8 @@ import pytest
 from langchain_core.embeddings import Embeddings
 
 from ontocast.config import ChunkConfig
-from ontocast.tool.chunk.util import SENTENCE_SPLIT_REGEX, SemanticChunker
+from ontocast.tool.chunk.proposition import SENTENCE_SPLIT_REGEX
+from ontocast.tool.chunk.util import SemanticChunker
 
 
 class TestSemanticChunker: